refactor db/clients.js and routes/clients.js
