+++ server/server.js
@@ -3,18 +3,18 @@ const path = require('path')
 
 const server = express()
 
-const taskRoutes = require('./routes/tasks')
+const itemRoutes = require('./routes/items')
 const clientRoutes = require('./routes/clients')
 const invoiceRoutes = require('./routes/invoices')
 
 server.use(express.json())
 server.use(express.static(path.join(__dirname, 'public')))
 
-server.use('/api/v1/tasks', taskRoutes)
 server.use('/api/v1/clients', clientRoutes)
+server.use('/api/v1/items', itemRoutes)
 server.use('/api/v1/invoices', invoiceRoutes)
 
-// server.use('/api/v1/*', (req, res) => res.sendStatus(404))
+server.use('/api/v1/*', (req, res) => res.sendStatus(404))
 
 server.get('*', (req, res) => {
   res.sendFile(path.join(__dirname, 'public/index.html'))
